Fix remaining struct initialization errors after proto consolidation
- Added missing proto fields to QuantizationConfig initializers
- Fixed FastLanesScheme variant field requirements
- Added missing 'source' field to VectorRecord initializations
- Fixed DataBlockMetadata, FlushResult, and CompactionResult missing fields
- Added encoding_marker and encoding_metadata to DataBlock structs
- Updated SearchPlan with available_quantization field

Reduced compilation errors from 405 to 385 (20 errors fixed)

+++ src/storage/engines/core/formats/columnar/hybrid_writer.rs
@@ -597,7 +597,7 @@ impl HybridParquetWriter {
         if flushes == 0 {
             self.stats.avg_flush_latency_ms.store(new_latency, Ordering::Relaxed);
         } else {
-            let new_avg = (current * (flushes - 1) + new_latency) / flushes;
+            let new_avg = (current * (flushes as u64 - 1) + new_latency) / flushes as u64;
             self.stats.avg_flush_latency_ms.store(new_avg, Ordering::Relaxed);
         }
     }
